requirements.txt
numpy==1.26.4
lightning.pytorch==2.3.3
PyYAML==6.0.1
torch==2.3.1
torchvision==0.18.1
pytorchvideo==0.1.5
